models: Support SubjectAltName to sign IP addresses
This adds support for a new table called subjectaltnames for lack of a better
name.

Subject Alternative Names are the new primary way of identifying a certificate,
as the Subject line is deprecated. This version of caramel will just copy the
commonname from the request,  and add it as a DNS SAN.

Next commit will add support for IP addresses to be added to SAN in the caramel
tool (not from the request itself). This allows us to further sign certain
certificates as belonging to IP-addresses, which is needed for Kubernetes

Issue: ModioAB/caramel#44